fix(as-3842): pr feedback
