+++ packages/forms-web-app/__tests__/unit/routes/before-you-start/any-of-following.test.js
@@ -1,5 +1,7 @@
-const { get, post } = require('../router-mock');
-const beforeYouStartController = require('../../../../src/controllers/before-you-start/any-of-following');
+const { get } = require('../router-mock');
+const {
+  getAnyOfFollowing,
+} = require('../../../../src/controllers/before-you-start/any-of-following');
 
 describe('routes/before-you-start/any-of-following', () => {
   beforeEach(() => {
@@ -12,7 +14,6 @@ describe('routes/before-you-start/any-of-following', () => {
   });
 
   it('should define the expected routes', () => {
-    expect(get).toHaveBeenCalledWith('/', beforeYouStartController.getAnyOfFollowing);
-    expect(post).toHaveBeenCalledWith('/', beforeYouStartController.postAnyOfFollowing);
+    expect(get).toHaveBeenCalledWith('/', getAnyOfFollowing);
   });
 });
