+++ packages/forms-web-app/__tests__/unit/routes/before-you-start/index.test.js
@@ -1,4 +1,6 @@
 const { use } = require('../router-mock');
+const anyOfFollowingRouter = require('../../../../src/routes/before-you-start/any-of-following');
+const localPlanningDepartmentRouter = require('../../../../src/routes/before-you-start/any-of-following');
 
 describe('routes/before-you-start/index', () => {
   beforeEach(() => {
@@ -13,6 +15,9 @@ describe('routes/before-you-start/index', () => {
   });
 
   it('should define the expected routes', () => {
+    expect(use).toHaveBeenCalledWith(localPlanningDepartmentRouter);
+    expect(use).toHaveBeenCalledWith('/any-of-following', anyOfFollowingRouter);
+
     expect(use.mock.calls.length).toBe(2);
   });
 });
