+++ packages/forms-web-app/src/routes/before-you-start/any-of-following.js
@@ -4,10 +4,13 @@ const {
   getAnyOfFollowing,
   postAnyOfFollowing,
 } = require('../../controllers/before-you-start/any-of-following');
+const {
+  rules: anyOfFollowingRules,
+} = require('../../validators/before-you-start/any-of-following');
 
 const router = express.Router();
 
 router.get('/', getAnyOfFollowing);
-router.post('/', postAnyOfFollowing);
+router.post('/', anyOfFollowingRules(), postAnyOfFollowing);
 
 module.exports = router;
