+++ packages/forms-web-app/src/routes/before-you-start/index.js
@@ -1,13 +1,10 @@
 const express = require('express');
 const anyOfFollowingRouter = require('./any-of-following');
 const localPlanningDepartmentRouter = require('./local-planning-department');
-const {
-  rules: anyOfFollowingRules,
-} = require('../../validators/before-you-start/any-of-following');
 
 const router = express.Router();
 
 router.use(localPlanningDepartmentRouter);
-router.use('/any-of-following', anyOfFollowingRules(), anyOfFollowingRouter);
+router.use('/any-of-following', anyOfFollowingRouter);
 
 module.exports = router;
