+++ packages/forms-web-app/src/views/before-you-start/any-of-following.njk
@@ -77,7 +77,7 @@
 
         {{ govukButton({
             text: "Continue",
-            type: "Submit",
+            type: "submit",
             attributes: { "data-cy":"button-save-and-continue"}
         }) }}
     </form>
